+++ src/ParticlesObstacles.jl
@@ -3,8 +3,8 @@ import Base: show, eltype, getindex
 
 export AbstractParticle, Particle, MagneticParticle,
 cyclotron, Obstacle, Disk, Antidot, RandomDisk,
-FiniteWall, PeriodicWall, RandomWall, SplitterWall,
-normalvec, randominside, distance, cellsize, Wall, Circular
+InfiniteWall, PeriodicWall, RandomWall, SplitterWall, FiniteWall,
+normalvec, randominside, distance, cellsize, Wall, Circular, Semicircle
 
 ####################################################
 ## Particles
@@ -181,7 +181,7 @@ RandomDisk{T}(args...) where {T} = RandomDisk(args...)
     Antidot{T<:AbstractFloat} <: Circular{T}
 Disk-like obstacle that allows propagation both inside and outside of the disk
 (immutable type). Used in ray-splitting billiards.
-#### Fields:
+### Fields:
 * `c::SVector{2,T}` : Center.
 * `r::T` : Radius.
 * `pflag::Bool` : Flag that keeps track of where the particle is currently
@@ -205,9 +205,34 @@ end
 Antidot(c, r, name::String = "Antidot") = Antidot(c, r, true, name)
 Antidot{T}(args...) where {T} = Antidot(args...)
 
+"""
+    Semicircle{T<:AbstractFloat} <: Circular{T}
+Obstacle that represents half a circle.
+### Fields:
+* `c::SVector{2,T}` : Center.
+* `r::T` : Radius.
+* `facedir::SVector{2,T}` : Direction where the open face of the Semicircle is facing.
+* `name::String` : Name of the obstacle given for user convenience.
+  Defaults to "Semicircle".
+"""
+struct Semicircle{T<:AbstractFloat} <: Circular{T}
+    c::SVector{2,T}
+    r::T
+    facedir::SVector{2,T}
+    name::String
+end
+function Semicircle(
+    c::AbstractVector{T}, r::Real, facedir, name = "Semicircle") where {T<:Real}
+    S = T <: Integer ? Float64 : T
+    return Semicircle{S}(
+    SV(c), convert(S, abs(r)), SV(normalize(facedir)), name)
+end
+
 show(io::IO, w::Circular{T}) where {T} =
 print(io, "$(w.name) {$T}\n", "center: $(w.c)\nradius: $(w.r)")
 
+show(io::IO, w::Semicircle{T}) where {T} =
+print(io, "$(w.name) {$T}\n", "center: $(w.c)\nradius: $(w.r)\nfacedir: $(w.facedir)")
 
 """
     Wall{T<:AbstractFloat} <: Obstacle{T}
@@ -216,8 +241,9 @@ Wall obstacle supertype.
 abstract type Wall{T<:AbstractFloat} <: Obstacle{T} end
 
 """
-    FiniteWall{T<:AbstractFloat} <: Wall{T}
+    InfiniteWall{T<:AbstractFloat} <: Wall{T}
 Wall obstacle imposing specular reflection during collision (immutable type).
+Faster than [`FiniteWall`](@ref), meant to be used for convex billiards.
 ### Fields:
 * `sp::SVector{2,T}` : Starting point of the Wall.
 * `ep::SVector{2,T}` : Ending point of the Wall.
@@ -228,13 +254,13 @@ Wall obstacle imposing specular reflection during collision (immutable type).
 * `name::String` : Name of the obstacle, given for user convenience.
   Defaults to "Wall".
 """
-struct FiniteWall{T<:AbstractFloat} <: Wall{T}
+struct InfiniteWall{T<:AbstractFloat} <: Wall{T}
     sp::SVector{2,T}
     ep::SVector{2,T}
     normal::SVector{2,T}
     name::String
 end
-function FiniteWall(sp::AbstractVector, ep::AbstractVector,
+function InfiniteWall(sp::AbstractVector, ep::AbstractVector,
     n::AbstractVector, name::String = "Wall")
     T = eltype(sp)
     n = normalize(n)
@@ -243,13 +269,60 @@ function FiniteWall(sp::AbstractVector, ep::AbstractVector,
         error("Normal vector is not actually normal to the wall")
     end
     T = eltype(sp) <: Integer ? Float64 : eltype(sp)
-    return FiniteWall{T}(SVector{2,T}(sp), SVector{2,T}(ep), SVector{2,T}(n), name)
+    return InfiniteWall{T}(SVector{2,T}(sp), SVector{2,T}(ep), SVector{2,T}(n), name)
 end
+
+"""
+    FiniteWall{T<:AbstractFloat} <: Wall{T}
+Wall obstacle imposing specular reflection during collision (immutable type).
+Slower than [`InfiniteWall`](@ref), meant to be used for non-convex billiards.
+
+Giving a `true` value to the field `isdoor` designates this obstacle to be a `Door`.
+This is used in [`escapetime`](@ref) function. A `Door` is a obstacle of the
+billiard table that the particle can escape from, thus enabling calculations
+of escape times.
+
+### Fields:
+* `sp::SVector{2,T}` : Starting point of the Wall.
+* `ep::SVector{2,T}` : Ending point of the Wall.
+* `normal::SVector{2,T}` : Normal vector to the wall, pointing to where the
+  particle *will come from before a collision* (pointing towards the inside of the
+  billiard table). The size of the vector is irrelevant
+  since it is internally normalized.
+* `isdoor::Bool` : Flag of whether this `FiniteWall` instance is a "Door".
+* `name::String` : Name of the obstacle, given for user convenience.
+  Defaults to "Finite Wall".
+"""
+struct FiniteWall{T<:AbstractFloat} <: Wall{T}
+    sp::SVector{2,T}
+    ep::SVector{2,T}
+    normal::SVector{2,T}
+    width::T
+    center::SVector{2,T}
+    isdoor::Bool
+    name::String
+end
+function FiniteWall(sp::AbstractVector, ep::AbstractVector,
+    n::AbstractVector, isdoor::Bool = false, name::String = "Finite Wall")
+    T = eltype(sp)
+    n = normalize(n)
+    d = dot(n, ep-sp)
+    if abs(d) > 10eps(T)
+        error("Normal vector is not actually normal to the wall: dot = $d")
+    end
+    T = eltype(sp) <: Integer ? Float64 : eltype(sp)
+    w = norm(ep - sp)
+    center = @. (ep+sp)/2
+    return FiniteWall{T}(SVector{2,T}(sp), SVector{2,T}(ep), SVector{2,T}(n),
+    w, SVector{2,T}(center), isdoor, name)
+end
+
+isdoor(w) = w.isdoor
 
 """
     RandomWall{T<:AbstractFloat} <: Wall{T}
 Wall obstacle imposing (uniformly) random reflection during collision (immutable type).
-#### Fields:
+### Fields:
 * `sp::SVector{2,T}` : Starting point of the Wall.
 * `ep::SVector{2,T}` : Ending point of the Wall.
 * `normal::SVector{2,T}` : Normal vector to the wall, pointing to where the
@@ -340,17 +413,17 @@ function SplitterWall(sp::AbstractVector, ep::AbstractVector,
         error("Normal vector is not actually normal to the wall")
     end
     T = eltype(sp) <: Integer ? Float64 : eltype(sp)
-    return SplitterWall{T}(SVector{2,T}(sp), SVector{2,T}(ep), SVector{2,T}(n), pflag, name)
+    return SplitterWall{T}(
+    SVector{2,T}(sp), SVector{2,T}(ep), SVector{2,T}(n), pflag, name)
 end
-SplitterWall(sp, ep, n, name::String = "Splitter wall") = SplitterWall(sp, ep, n, true, name)
+SplitterWall(sp, ep, n, name::String = "Splitter wall") =
+SplitterWall(sp, ep, n, true, name)
 #pretty print:
 show(io::IO, w::Wall{T}) where {T} = print(io, "$(w.name) {$T}\n",
 "start point: $(w.sp)\nend point: $(w.ep)\nnormal vector: $(w.normal)")
 
 """
-```julia
-normalvec(obst::Obstacle, position)
-```
+    normalvec(obst::Obstacle, position)
 Return the vector normal to the obstacle's boundary at the given position (which is
 assumed to be very close to the obstacle's boundary).
 """
@@ -359,15 +432,14 @@ normalvec(w::PeriodicWall, pos) = normalize(w.normal)
 normalvec(w::SplitterWall, pos) = w.pflag ? w.normal : -w.normal
 normalvec(disk::Circular, pos) = normalize(pos - disk.c)
 normalvec(a::Antidot, pos) = a.pflag ? normalize(pos - a.c) : -normalize(pos - a.c)
+normalvec(d::Semicircle, pos) = normalize(d.c - pos)
 
 ####################################################
 ## Billiard Table
 ####################################################
 function isperiodic(bt)::Bool
     for obst in bt
-        if typeof(obst) <: PeriodicWall
-            return true
-        end
+        typeof(obst) <: PeriodicWall && return true
     end
     return false
 end
@@ -415,23 +487,23 @@ the billiard table.
 
 All `distance` functions can also be given a position (Vector) instead of a particle.
 """
-function distance(pos::AbstractVector{T}, v::Vector{<:Obstacle{T}})::T where {T}
+function distance(p::AbstractParticle{T}, v::Vector{<:Obstacle{T}})::T where {T}
     d = T(Inf)
     for obst in v
-        di = distance(pos, obst)
+        di = distance(p, obst)
         di < d && (d = di)
     end
     return d
 end
 
-(distance(p::AbstractParticle{T}, v::Vector{<:Obstacle{T}})::T) where {T} =
-distance(p.pos, v)
+(distance(p::AbstractParticle{T}, obst::Obstacle{T})::T) where {T} =
+distance(p.pos, obst)
 
-(distance(pos::AbstractVector{T}, bt::Tuple)::T) where {T<:AbstractFloat} =
-min(distance(pos, obst) for obst in bt)
-
-distance(pos::AbstractVector, bt::BilliardTable) = distance(pos, bt.bt)
-distance(p::AbstractParticle, bt::BilliardTable) = distance(p.pos, bt.bt)
+# (distance(pos::AbstractVector{T}, bt::Tuple)::T) where {T<:AbstractFloat} =
+# min(distance(pos, obst) for obst in bt)
+#
+# distance(pos::AbstractVector, bt::BilliardTable) = distance(pos, bt.bt)
+# distance(p::AbstractParticle, bt::BilliardTable) = distance(p.pos, bt.bt)
 
 function distance(pos::AbstractVector{T}, w::Wall{T})::T where {T}
     v1 = pos - w.sp
@@ -451,10 +523,65 @@ function distance(
 end
 distance(pos::AbstractVector{T}, a::Antidot{T}) where {T} = distance(pos, a, a.pflag)
 
-(distance(p::AbstractParticle{T}, obst::Obstacle{T})::T) where {T} =
-distance(p.pos, obst)
+function distance(pos::AbstractVector{T}, s::Semicircle{T}) where {T}
+    # Check on which half of circle is the particle
+    v1 = pos .- s.c
+    nn = dot(v1, s.facedir)
+    if nn ≤ 0 # I am "inside semicircle
+        return s.r - norm(pos - s.c)
+    else # I am on the "other side"
+        end1 = SV(s.c[1] + s.r*s.facedir[2], s.c[2] - s.r*s.facedir[1])
+        end2 = SV(s.c[1] - s.r*s.facedir[2], s.c[2] + s.r*s.facedir[1])
+        return min(norm(pos - end1), norm(pos - end2))
+    end
+end
 
 
+
+# Distances for randominside:
+distance_init(p, a) = distance(p, a)
+
+function distance_init(p::AbstractParticle{T}, v::Vector{<:Obstacle{T}})::T where {T}
+    d = T(Inf)
+    for obst in v
+        di = distance_init(p, obst)
+        di < d && (d = di)
+    end
+    return d
+end
+
+function distance_init(p::AbstractParticle{T}, w::FiniteWall{T})::T where {T}
+
+    n = normalvec(w, p.pos)
+    posdot = dot(w.sp - p.pos, n)
+    if posdot ≥ 0 # I am behind wall
+        intersection = project_to_line(p.pos, w.center, n)
+        dfc = norm(intersection - w.center)
+        if dfc > w.width/2
+            return +1 # but not directly behind
+        else
+            return -1
+        end
+    end
+    v1 = p.pos - w.sp
+    dot(v1, n)
+end
+
+"""
+    project_to_line(point, c, n)
+Project given `point` to line that contains point `c` and has **normal vector** `n`.
+Return the projected point.
+"""
+function project_to_line(point, c, n)
+    posdot = dot(c - point, n)
+    intersection = point .+ posdot.* n
+end
+# function project_to_line(point, c, n)
+#     posdot = dot(c - point, n)
+#     denom = dot(n, n)
+#     colt = posdot/denom
+#     intersection = point .+ colt .* n
+# end
 ####################################################
 ## Initial Conditions
 ####################################################
@@ -487,6 +614,12 @@ function cellsize(a::Antidot{T}) where {T}
         xmin, ymin = a.c .- a.r
         xmax, ymax = a.c .+ a.r
     end
+    return xmin, ymin, xmax, ymax
+end
+
+function cellsize(a::Semicircle{T}) where {T}
+    xmin, ymin = a.c .- a.r
+    xmax, ymax = a.c .+ a.r
     return xmin, ymin, xmax, ymax
 end
 
@@ -524,13 +657,13 @@ function randominside(bt::Vector{<:Obstacle{T}}) where {T<:AbstractFloat}
     yp = T(rand())*(ymax-ymin) + ymin
     p = Particle([xp, yp, φ0])
 
-    dist = distance(p, bt)
+    dist = distance_init(p, bt)
     while dist <= sqrt(eps(T))
 
         xp = T(rand())*(xmax-xmin) + xmin
         yp = T(rand())*(ymax-ymin) + ymin
         p.pos = SVector{2,T}(xp, yp)
-        dist = distance(p, bt)
+        dist = distance_init(p, bt)
     end
 
     return p
@@ -553,13 +686,13 @@ function randominside(ω::Real, bt::Vector{<:Obstacle{T}}) where {T<:AbstractFlo
     yp = T(rand())*(ymax-ymin) + ymin
     p = MagneticParticle([xp, yp, φ0], ω)
 
-    dist = distance(p, bt)
+    dist = distance_init(p, bt)
     while dist <= sqrt(eps(T))
 
         xp = rand()*(xmax-xmin) + xmin
         yp = rand()*(ymax-ymin) + ymin
         p.pos = [xp, yp]
-        dist = distance(p, bt)
+        dist = distance_init(p, bt)
     end
 
     return p